Moved spartan CNI functionality into a separate plugin package.
This will allow us to split the spartan functionality into its own
plugin once we have service chaining in Mesos.
